reduced CoffeeShop Decorator exercise to fewer files
