--- Design-Patterns/Decorator-Pattern/Beverages/DarkRoast.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Decorator_Pattern.Beverages
-{
-    internal class DarkRoast : Beverage
-    {
-        public DarkRoast() : base("A Dark Roast")
-        {
-        }
-
-        public override decimal Cost()
-        {
-            return 3.10M;
-        }
-    }
-}
